+++ Functions/Import and Export/excel_Callback.m
@@ -1,7 +1,7 @@
 function excel_Callback(hObject, eventdata, handles)
 
     function t = loop_calls(Calls, hc,includereject,waitbar_text,handles,call_file, audioReader)
-            exceltable = [{'ID'} {'Label'} {'Accepted'} {'Score'} {'Begin Time (s)'} {'End Time (s)'} {'Call Length (s)'} {'Principal Frequency (kHz)'} {'Low Freq (kHz)'} {'High Freq (kHz)'} {'Delta Freq (kHz)'} {'Frequency Standard Deviation (kHz)'} {'Slope (kHz/s)'} {'Sinuosity'} {'Mean Power (dB/Hz)'} {'Tonality'} {'Peak Freq (kHz)'}];        for i = 1:height(Calls) % Do this for each call
+            exceltable = [{'File'} {'ID'} {'Label'} {'Accepted'} {'Score'} {'Begin Time (s)'} {'End Time (s)'} {'Call Length (s)'} {'Principal Frequency (kHz)'} {'Low Freq (kHz)'} {'High Freq (kHz)'} {'Delta Freq (kHz)'} {'Frequency Standard Deviation (kHz)'} {'Slope (kHz/s)'} {'Sinuosity'} {'Mean Power (dB/Hz)'} {'Tonality'} {'Peak Freq (kHz)'}];        for i = 1:height(Calls) % Do this for each call
             waitbar(i/height(Calls),hc,waitbar_text);
 
             if includereject || Calls.Accept(i)
@@ -24,7 +24,8 @@ function excel_Callback(hObject, eventdata, handles)
                 Label = Calls.Type(i);
                 Score = Calls.Score(i);
                 accepted = Calls.Accept(i);
-                exceltable = [exceltable; {ID} {Label} {accepted} {Score} {stats.BeginTime} {stats.EndTime} {stats.DeltaTime} {stats.PrincipalFreq} {stats.LowFreq} {stats.HighFreq} {stats.DeltaFreq} {stats.stdev} {stats.Slope} {stats.Sinuosity} {stats.MeanPower} {stats.SignalToNoise} {stats.PeakFreq}];            end
+                File = call_file;
+                exceltable = [exceltable; {File} {ID} {Label} {accepted} {Score} {stats.BeginTime} {stats.EndTime} {stats.DeltaTime} {stats.PrincipalFreq} {stats.LowFreq} {stats.HighFreq} {stats.DeltaFreq} {stats.stdev} {stats.Slope} {stats.Sinuosity} {stats.MeanPower} {stats.SignalToNoise} {stats.PeakFreq}];            end
 
         end
         t = cell2table(exceltable);